+++ Project/src/Instructions/MIPSInstructionR.java
@@ -30,13 +30,13 @@ public class MIPSInstructionR extends AbstractMIPSInstruction {
             shamt = 0;
             funct = FunctUtil.getValue(arr[0]);
         } else if(type == MIPSStringType.Hex || type == MIPSStringType.Binary) {
-            int val = getValue(str, type);
+            long val = getValue(str, type);
             op = 0;
-            rs = (val & ((1 << 26) - 1)) >> 21;
-            rt = (val & ((1 << 21) - 1)) >> 16;
-            rd = (val & ((1 << 16) - 1)) >> 11;
+            rs = (int)((val & ((1 << 26) - 1)) >> 21);
+            rt = (int)((val & ((1 << 21) - 1)) >> 16);
+            rd = (int)((val & ((1 << 16) - 1)) >> 11);
             shamt = 0;
-            funct = val & ((1 << 6) - 1);
+            funct = (int)(val & ((1 << 6) - 1));
         } else {
             throw new UnsupportedOperationException("MIPSString type not implemented yet");
         }
